adjust：调整Controller接口分布，调整Redis配置
