+++ src/main/java/com/luckybird/springbackend/controller/UserController.java
@@ -1,11 +1,12 @@
 package com.luckybird.springbackend.controller;
 
 import com.luckybird.springbackend.api.UserApi;
-import com.luckybird.springbackend.api.req.UserCreateReq;
-import com.luckybird.springbackend.api.req.UserQueryReq;
-import com.luckybird.springbackend.api.req.UserUpdateReq;
+import com.luckybird.springbackend.api.req.*;
 import com.luckybird.springbackend.api.vo.UserVO;
+import com.luckybird.springbackend.common.annotation.NoAuth;
 import com.luckybird.springbackend.common.base.PageResult;
+import com.luckybird.springbackend.common.base.TokenInfo;
+import com.luckybird.springbackend.common.util.ContextUtil;
 import com.luckybird.springbackend.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -117,4 +118,35 @@ public class UserController implements UserApi {
             @RequestBody UserQueryReq req) {
         return userService.page(req, current, rows, searchCount);
     }
+
+    /**
+     * 登录用户
+     *
+     * @param req 用户登录请求
+     * @return UserVO 用户信息
+     */
+    @PostMapping("/v1/users/login")
+    @NoAuth
+    public TokenInfo login(@RequestBody @Valid UserLoginReq req) {
+        return userService.login(req);
+    }
+
+    /**
+     * 登出用户
+     */
+    @PostMapping("/v1/users/logout")
+    public void logout() {
+        userService.logout(ContextUtil.getUserInfo().getId());
+    }
+
+    /**
+     * 修改密码
+     *
+     * @param req 密码修改请求
+     */
+    @PostMapping("/v1/users/changePassword")
+    public void changePassword(@RequestBody @Valid UserChangePasswordReq req) {
+        userService.changePassword(ContextUtil.getUserInfo().getId(), req);
+    }
+
 }
